+++ test/02-randombytes.quick.js
@@ -2,11 +2,11 @@ var nacl = (typeof window !== 'undefined') ? window.nacl : require('../' + (proc
 nacl.util = require('tweetnacl-util');
 var test = require('tape');
 
-test('nacl.randomBytes', function(t) {
+test('nacl.randomBytes', async function(t) {
   t.plan(1);
   var set = {}, s, i;
   for (i = 0; i < 10000; i++) {
-    s = nacl.util.encodeBase64(nacl.randomBytes(32));
+    s = nacl.util.encodeBase64(await nacl.randomBytes(32));
     if (set[s]) {
       t.fail('duplicate random sequence! ', s);
       return;