+++ test/06-box.quick.js
@@ -4,25 +4,25 @@ var test = require('tape');
 
 var enc = nacl.util.encodeBase64;
 
-test('nacl.box.keyPair', function(t) {
-  var keys = nacl.box.keyPair();
+test('nacl.box.keyPair', async function(t) {
+  var keys = await nacl.box.keyPair();
   t.ok(keys.secretKey && keys.secretKey.length === nacl.box.secretKeyLength, 'has secret key');
   t.ok(keys.publicKey && keys.publicKey.length === nacl.box.publicKeyLength, 'has public key');
   t.notEqual(enc(keys.secretKey), enc(keys.publicKey));
   t.end();
 });
 
-test('nacl.box.keyPair.fromSecretKey', function(t) {
-  var k1 = nacl.box.keyPair();
+test('nacl.box.keyPair.fromSecretKey', async function(t) {
+  var k1 = await nacl.box.keyPair();
   var k2 = nacl.box.keyPair.fromSecretKey(k1.secretKey);
   t.equal(enc(k2.secretKey), enc(k1.secretKey));
   t.equal(enc(k2.publicKey), enc(k1.publicKey));
   t.end();
 });
 
-test('nacl.box and nacl.box.open', function(t) {
-  var clientKeys = nacl.box.keyPair();
-  var serverKeys = nacl.box.keyPair();
+test('nacl.box and nacl.box.open', async function(t) {
+  var clientKeys = await nacl.box.keyPair();
+  var serverKeys = await nacl.box.keyPair();
   var nonce = new Uint8Array(nacl.box.nonceLength);
   for (var i = 0; i < nonce.length; i++) nonce[i] = (32+i) & 0xff;
   var msg = nacl.util.decodeUTF8('message to encrypt');
@@ -36,9 +36,9 @@ test('nacl.box and nacl.box.open', function(t) {
   t.end();
 });
 
-test('nacl.box.open with invalid box', function(t) {
-  var clientKeys = nacl.box.keyPair();
-  var serverKeys = nacl.box.keyPair();
+test('nacl.box.open with invalid box', async function(t) {
+  var clientKeys = await nacl.box.keyPair();
+  var serverKeys = await nacl.box.keyPair();
   var nonce = new Uint8Array(nacl.box.nonceLength);
   t.equal(nacl.box.open(new Uint8Array(0), nonce, serverKeys.publicKey, clientKeys.secretKey), null);
   t.equal(nacl.box.open(new Uint8Array(10), nonce, serverKeys.publicKey, clientKeys.secretKey), null);
@@ -46,9 +46,9 @@ test('nacl.box.open with invalid box', function(t) {
   t.end();
 });
 
-test('nacl.box.open with invalid nonce', function(t) {
-  var clientKeys = nacl.box.keyPair();
-  var serverKeys = nacl.box.keyPair();
+test('nacl.box.open with invalid nonce', async function(t) {
+  var clientKeys = await nacl.box.keyPair();
+  var serverKeys = await nacl.box.keyPair();
   var nonce = new Uint8Array(nacl.box.nonceLength);
   for (var i = 0; i < nonce.length; i++) nonce[i] = i & 0xff;
   var msg = nacl.util.decodeUTF8('message to encrypt');
@@ -60,9 +60,9 @@ test('nacl.box.open with invalid nonce', function(t) {
   t.end();
 });
 
-test('nacl.box.open with invalid keys', function(t) {
-  var clientKeys = nacl.box.keyPair();
-  var serverKeys = nacl.box.keyPair();
+test('nacl.box.open with invalid keys', async function(t) {
+  var clientKeys = await nacl.box.keyPair();
+  var serverKeys = await nacl.box.keyPair();
   var nonce = new Uint8Array(nacl.box.nonceLength);
   var msg = nacl.util.decodeUTF8('message to encrypt');
   var box = nacl.box(msg, nonce, clientKeys.publicKey, serverKeys.secretKey);